Create the component incrementer
