+++ src/app/components/incrementador/incrementador.component.html
@@ -1,0 +1,20 @@
+<div class="input-group">
+    <span class="input-group-btn">
+        <button class="btn btn-primary" 
+                (click)="cambiarValor(-5)"
+                type="button">
+            <i class="fa fa-minus"></i>
+        </button>
+    </span>
+    <input type="text" 
+           class="form-control" 
+           placeholder="Progreso"
+           [(ngModel)]="progreso">
+    <span class="input-group-btn">
+        <button class="btn btn-primary" 
+                (click)="cambiarValor(5)"
+                type="button">
+            <i class="fa fa-plus"></i>
+        </button>
+    </span>
+</div>
