+++ src/app/components/incrementador/incrementador.component.ts
@@ -1,0 +1,24 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-incrementador',
+  templateUrl: './incrementador.component.html'
+})
+export class IncrementadorComponent {
+   progreso: number = 50;
+
+  get getPorcentaje() {
+    return `${ this.progreso }%`;
+  }
+
+  cambiarValor(valor: number){
+
+    if( this.progreso >= 100 && valor >= 0){
+     return this.progreso = 100;
+    }
+    if( this.progreso <= 0 && valor < 0){
+      return this.progreso = 0;
+     }
+   return  this.progreso = this.progreso + valor;
+  }
+}
