+++ src/app/pages/progress/progress.component.html
@@ -6,10 +6,19 @@
                 <div class="progress m-t-20">
                     <div class="progress-bar bg-primary progress-bar-striped active"  
                         style="height:5px;" 
-                        [style.width]="getPorcentaje"
                         role="progressbar"> 
+                         <!--- [style.width]="getPorcentaje" -->
                     </div>
                 </div>
+
+                <div class="progress m-t-20">
+                    <div class="progress-bar bg-info progress-bar-striped active"  
+                        style="height:5px;" 
+                        role="progressbar"> 
+                           <!--- [style.width]="getPorcentaje" -->
+                    </div>
+                </div>
+
             </div>
         </div>
     </div>
@@ -18,26 +27,7 @@
     <div class="col-6">
         <div class="card">
             <div class="card-body">
-                <div class="input-group">
-                    <span class="input-group-btn">
-                        <button class="btn btn-primary" 
-                                (click)="cambiarValor(-5)"
-                                type="button">
-                            <i class="fa fa-minus"></i>
-                        </button>
-                    </span>
-                    <input type="text" 
-                           class="form-control" 
-                           placeholder="Progreso"
-                           [(ngModel)]="progreso">
-                    <span class="input-group-btn">
-                        <button class="btn btn-primary" 
-                                (click)="cambiarValor(5)"
-                                type="button">
-                            <i class="fa fa-plus"></i>
-                        </button>
-                    </span>
-                </div>
+                <app-incrementador></app-incrementador>
             </div>
         </div>
 
@@ -45,7 +35,7 @@
     <div class="col-6">
         <div class="card">
             <div class="card-body">
-                2
+                <app-incrementador></app-incrementador>
             </div>
         </div>
