+++ src/app/pages/progress/progress.component.ts
@@ -6,20 +6,5 @@ import { Component } from '@angular/core';
   styleUrls:['./progress.component.css']
 })
 export class ProgressComponent {
-  progreso: number = 50;
-
-  get getPorcentaje() {
-    return `${ this.progreso }%`;
-  }
-
-  cambiarValor(valor: number){
-
-    if( this.progreso >= 100 && valor >= 0){
-     return this.progreso = 100;
-    }
-    if( this.progreso <= 0 && valor < 0){
-      return this.progreso = 0;
-     }
-   return  this.progreso = this.progreso + valor;
-  }
+  
 }
